+++ src/onegov.election_day/onegov/election_day/formats/election/wabsti/majorz/__init__.py
@@ -1,0 +1,10 @@
+from onegov.election_day.formats.election.wabsti.majorz.simple import \
+    import_file
+from onegov.election_day.formats.election.wabsti.majorz.exporter import \
+    import_exporter_files
+
+
+__all__ = [
+    'import_file',
+    'import_exporter_files'
+]