+++ src/onegov.election_day/onegov/election_day/forms/election.py
@@ -66,6 +66,45 @@ class ElectionForm(Form):
         label=_("Related link")
     )
 
+    upload_type = RadioField(
+        _("Upload"),
+        choices=[
+            ('manual', _("Manual")),
+            ('wabsti', _("Automatic (Wabsti)")),
+        ],
+        validators=[
+            InputRequired()
+        ],
+        default='manual'
+    )
+
+    upload_wabsti_district = StringField(
+        label=_("Automatic upload (Wabsti): 'SortWahlkreis'"),
+        validators=[
+            InputRequired()
+        ],
+        render_kw=dict(force_simple=True),
+        depends_on=('upload_type', 'wabsti'),
+    )
+
+    upload_wabsti_number = StringField(
+        label=_("Automatic upload (Wabsti): 'SortGeschaeft'"),
+        validators=[
+            InputRequired()
+        ],
+        render_kw=dict(force_simple=True),
+        depends_on=('upload_type', 'wabsti'),
+    )
+
+    upload_wabsti_folder = StringField(
+        label=_("Automatic upload (Wabsti): Folder"),
+        validators=[
+            InputRequired()
+        ],
+        render_kw=dict(force_simple=True),
+        depends_on=('upload_type', 'wabsti')
+    )
+
     def set_domain(self, principal):
         self.domain.choices = [
             (key, text)
@@ -94,6 +133,10 @@ class ElectionForm(Form):
         if not model.meta:
             model.meta = {}
         model.meta['related_link'] = self.related_link.data
+        model.meta['upload_type'] = self.upload_type.data
+        model.meta['upload_wabsti_district'] = self.upload_wabsti_district.data
+        model.meta['upload_wabsti_number'] = self.upload_wabsti_number.data
+        model.meta['upload_wabsti_folder'] = self.upload_wabsti_folder.data
 
     def apply_model(self, model):
         self.election_de.data = model.title_translations['de_CH']
@@ -109,3 +152,13 @@ class ElectionForm(Form):
 
         meta_data = model.meta or {}
         self.related_link.data = meta_data.get('related_link', '')
+        self.upload_type.data = meta_data.get('upload_type', 'manual')
+        self.upload_wabsti_district.data = meta_data.get(
+            'upload_wabsti_district', ''
+        )
+        self.upload_wabsti_number.data = meta_data.get(
+            'upload_wabsti_number', ''
+        )
+        self.upload_wabsti_folder.data = meta_data.get(
+            'upload_wabsti_folder', ''
+        )