+++ src/onegov.election_day/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Shows the filename of the import errors.
   [msom]
 
+- Adds support for automatic import of votes using the Wabsti format.
+  [msom]
+
 1.7.5 (2017-04-07)
 ~~~~~~~~~~~~~~~~~~~